+++ de.bund.bfr.knime.fsklab.nodes/src/de/bund/bfr/knime/fsklab/nodes/EditorNodeDialog.java
@@ -21,6 +21,7 @@ package de.bund.bfr.knime.fsklab.nodes;
 import java.awt.BorderLayout;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
+import java.awt.event.ItemEvent;
 import java.io.File;
 import java.io.IOException;
 import java.net.MalformedURLException;
@@ -67,13 +68,14 @@ import org.knime.core.node.workflow.NodeContext;
 import org.knime.core.node.workflow.WorkflowContext;
 import org.knime.core.util.FileUtil;
 import de.bund.bfr.knime.fsklab.FskPortObject;
+import de.bund.bfr.knime.fsklab.nodes.EditorNodeSettings.ModelType;
 import de.bund.bfr.knime.fsklab.nodes.common.ui.FBrowseButton;
 
-class FSKEditorJSNodeDialog extends DataAwareNodeDialogPane {
+class EditorNodeDialog extends DataAwareNodeDialogPane {
 
-  private FSKEditorJSNodeSettings settings;
+  private EditorNodeSettings settings;
 
-  private final DefaultComboBoxModel<String> modeltype;
+  private final DefaultComboBoxModel<ModelType> modelType;
 
   private final FilesHistoryPanel m_readmePanel;
 
@@ -86,15 +88,11 @@ class FSKEditorJSNodeDialog extends DataAwareNodeDialogPane {
   private File currentWorkingDirectory;
 
   private WorkingDirectoryChangeListener changeListener = new WorkingDirectoryChangeListener();
+  
+  public EditorNodeDialog() {
+    settings = new EditorNodeSettings();
 
-  private static final String[] MODEL_TYPES = {"GenericModel", "DataModel", "PredictiveModel",
-      "ExposureModel", "ToxicologicalModel", "DoseResponseModel", "ProcessModel",
-      "ConsumptionModel", "HealthModel", "RiskModel", "QraModel", "OtherModel"};
-
-  public FSKEditorJSNodeDialog() {
-    settings = new FSKEditorJSNodeSettings();
-
-    modeltype = new DefaultComboBoxModel<>(MODEL_TYPES);
+    modelType = new DefaultComboBoxModel<>(ModelType.values());
 
     FlowVariableModel readmeVariable = createFlowVariableModel("readme", FlowVariable.Type.STRING);
     m_readmePanel =
@@ -116,7 +114,12 @@ class FSKEditorJSNodeDialog extends DataAwareNodeDialogPane {
       throws NotConfigurableException {
     try {
       this.settings.load(settings);
-      modeltype.setSelectedItem(this.settings.modelType);
+      
+      // Select the model type in settings
+      if (StringUtils.isNotEmpty(this.settings.modelType)) {
+        modelType.setSelectedItem(ModelType.valueOf(this.settings.modelType));
+      }
+      
       m_readmePanel.setSelectedFile(this.settings.getReadme());
       m_workingDirectoryPanel.setSelectedFile(this.settings.getWorkingDirectory());
 
@@ -149,7 +152,7 @@ class FSKEditorJSNodeDialog extends DataAwareNodeDialogPane {
   protected void loadSettingsFrom(NodeSettingsRO settings, PortObject[] input)
       throws NotConfigurableException {
 
-    final FSKEditorJSNodeSettings editorSettings = new FSKEditorJSNodeSettings();
+    final EditorNodeSettings editorSettings = new EditorNodeSettings();
     try {
       editorSettings.load(settings);
     } catch (InvalidSettingsException exception) {
@@ -196,7 +199,8 @@ class FSKEditorJSNodeDialog extends DataAwareNodeDialogPane {
     } catch (IOException | URISyntaxException e) {
       e.printStackTrace();
     }
-    this.settings.modelType = (String) modeltype.getSelectedItem();
+    
+    this.settings.modelType = ((ModelType) modelType.getSelectedItem()).name();
     this.settings = editorSettings;
   }
 
@@ -223,8 +227,12 @@ class FSKEditorJSNodeDialog extends DataAwareNodeDialogPane {
     modelTypePanel.setBorder(
         BorderFactory.createTitledBorder(BorderFactory.createEtchedBorder(), "Model Type:"));
 
-    JComboBox<String> combo = new JComboBox<>(modeltype);
-    combo.addItemListener(event -> settings.modelType = (String) event.getItem());
+    JComboBox<ModelType> combo = new JComboBox<>(modelType);
+    combo.addItemListener(event -> {
+      if (event.getStateChange() == ItemEvent.SELECTED) {
+        settings.modelType = ((ModelType) event.getItem()).name();
+      }
+    });
     modelTypePanel.add(combo, BorderLayout.NORTH);
 
     modelTypePanel.add(Box.createHorizontalGlue());